[Feat] 매수, 매도시 거래내역, 보유주식 반영 구현
매수, 매도할 때 시뮬 날짜 주식데이터와 비교하여 매도, 매수 수행 이후 보유 주식과 거래내역에 반영

Issue:none

+++ back/src/main/java/com/udteam/miristock/dto/StockDataRequestDto.java
@@ -14,9 +14,9 @@ public class StockDataRequestDto {
     private String stockCode;
     private String stockName;
     private Integer stockDataDate;
-    private Integer stockDataClosingPrice;
+    private Long stockDataClosingPrice;
     private Long stockDataAmount;
-    private Integer stockDataPriceIncreasement;
+    private Long stockDataPriceIncreasement;
     private Float stockDataFlucauationRate;
 
     public StockDataEntity toEntity(){
