+++ back/src/main/java/com/udteam/miristock/entity/StockDataEntity.java
@@ -28,13 +28,13 @@ public class StockDataEntity {
     private Integer stockDataDate;
 
     @Column(name = "stockdata_closing_price")
-    private Integer stockDataClosingPrice;
+    private Long stockDataClosingPrice;
 
     @Column(name = "stockdata_amount")
     private Long stockDataAmount;
 
     @Column(name = "stockdata_price_increasement")
-    private Integer stockDataPriceIncreasement;
+    private Long stockDataPriceIncreasement;
 
     @Column(name = "stockdata_flucauation_rate")
     private Float stockDataFlucauationRate;
